+++ contracts/smartAccounts/interface/IIReaderFacet.sol
@@ -1,0 +1,12 @@
+// SPDX-License-Identifier: MIT
+pragma solidity >=0.7.6 <0.9;
+
+import {IReaderFacet} from "../../userInterfaces/facets/IReaderFacet.sol";
+
+/**
+ * @title IIReaderFacet
+ * @notice Internal interface for the ReaderFacet contract.
+ */
+interface IIReaderFacet is IReaderFacet {
+
+}