+++ contracts/userInterfaces/IMasterAccountController.sol
@@ -13,7 +13,7 @@ import {IInstructionsFacet} from "./facets/IInstructionsFacet.sol";
 import {IMemoInstructionsFacet} from "./facets/IMemoInstructionsFacet.sol";
 import {IPaymentProofsFacet} from "./facets/IPaymentProofsFacet.sol";
 import {IPersonalAccountsFacet} from "./facets/IPersonalAccountsFacet.sol";
-import {IPersonalAccountReaderFacet} from "./facets/IPersonalAccountReaderFacet.sol";
+import {IReaderFacet} from "./facets/IReaderFacet.sol";
 import {ITimelockFacet} from "./facets/ITimelockFacet.sol";
 import {IVaultsFacet} from "./facets/IVaultsFacet.sol";
 import {IPauseFacet} from "./facets/IPauseFacet.sol";
@@ -37,7 +37,7 @@ interface IMasterAccountController is
     IPauseFacet,
     IPaymentProofsFacet,
     IPersonalAccountsFacet,
-    IPersonalAccountReaderFacet,
+    IReaderFacet,
     ITimelockFacet,
     IVaultsFacet,
     IXrplProviderWalletsFacet